// backend/gestion-livraison-de-colis-api/src/common/dto/livreur/scroring-classement-dto.ts
export class LivreurScoringClassement {
    idLivreur: number;
    classement: number;
    nbrLivraisonJour: number;
    totalLivraisonEffectueJour: number;
    nbrLivraisonSemaine: number;
    totalLivraisonEffectueSemaine: number;
    totalLivraisonGlobal: number;
    totalLivraisonEffectueGlobal: number;
}